Reintegrate legacy UMS.
As downstream still carries our old UMS driver in order to workaround
the catastrophic incoherency issues that currently plague the earlier
i8xx chipsets when trying to use GEM, it makes sense to integrate their
work into the main repository so that their fixes are shared and
available for all users.

This combines the EXA code from 2.6 with the most recent updates through
to 2.9 when the UMS code was removed and the distributions were forced
to fork.