Drop hashes in titles.
The cases where error codes are *not* present are generally due to
out-of-memory conditions, and so the bug reports are otherwise
unactionable.  So it doesn't matter if they accidentally get duped.

The case of a gpu lockup having a single error code is appearing to be
rare enough in practice that we don't need to special case that.